+++ express-controllers/controller.js
@@ -2,6 +2,7 @@ const { clusterRoutes } = require("./cluster-api");
 const { schematicsRoutes } = require("./schematics-api");
 const { vsiRoutes } = require("./vsi-api");
 const { powerRoutes } = require("./power-api");
+const { statsRoutes } = require("./stats-api");
 
 /**
  * controller constructor
@@ -19,7 +20,7 @@ function controller(axios) {
   this.vsiRoutes = vsiRoutes(axios, this);
   this.schematicsRoutes = schematicsRoutes(axios, this);
   this.powerRoutes = powerRoutes(axios, this);
-
+  this.statsRoutes = statsRoutes(axios, this);
   /**
    * check if a token is expired
    * @returns {boolean} true if expired