+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - Users can now view Power VS Resources with no selected workspace on the CRAIG V2 Power VS page
 - Users can now view Power VS Resources with no selected subnets on the CRAIG V2 Power VS page
 - When updating a Power VS Workspace name, Transit Gateway connections will now replace the existing name with the new one
+- Graphs of CRAIG usage statistics (clones, views, references) are now available at `/stats` and `/v2/stats`
 - When using CRAIG V2 an empty space will now display in subnet tiers where all three possible zones are not present
 - When fetching VSI Images, the CRAIG API will now respond with up to 100 results instead of the default of 50